chore: split jni and ffi out as separate features (#288)
We should not be unconditionally building FFI. We also should not be just building JNI on a per-target basis. Both of these should be behind feature flags.

This PR also makes the integration tests ignored, so that they get run in a separate pass and thus don't get invoked repeatedly by `cargo hack`.